Fix test compile failures
